+++ src/types.ts
@@ -59,7 +59,8 @@ export type ResolutionResponse = {
   };
   gundb?: {
     username?: string;
-  };
+    public_key?: string;
+  },
   addresses: {
     [key: string]: string;
   };