+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 * Added web3 providers supports such as http and ws protocols
 * Introduced ConfigurationError which handles all error that were made cause of falty configurations
 
+## 1.5.1
+* fixing the version 
+
+## 1.4.1
+* Resolution#chatPk -> get a gundb public key from domain's record
+* Fix the bug with Resolution#chatId for ens domains
+
 ## 1.4.0
 * Resolve custom records
 * Resolution#chatId -> get a gundb chat id from domain's record